Time: 74 ms (10.38%) | Memory: 39.5 MB (5.94%) - LeetSync

+++ 2090-number-of-ways-to-arrive-at-destination/number-of-ways-to-arrive-at-destination.cpp
@@ -1,5 +1,5 @@
 #define ll long long
-#define pll pair<long long,long long>
+#define pll pair<ll,ll>
 class Solution {
 public:
     int countPaths(int n, vector<vector<int>>& roads) {
